requirements.txt
streamlit
python-docx
PyPDF2
groq
python-dotenv